Refactor: campaignId from Long to Int

+++ WordPress/src/main/java/org/wordpress/android/ui/mysite/cards/blaze/BlazeCampaignsCard.kt
@@ -54,7 +54,7 @@ fun BlazeCampaignsCard(
                     modifier = Modifier
                         .padding(start = 16.dp, end = 16.dp, bottom = 8.dp)
                         .clickable {
-                            blazeCampaignCardModel.campaign.onClick(blazeCampaignCardModel.campaign.id.toInt())
+                            blazeCampaignCardModel.campaign.onClick(blazeCampaignCardModel.campaign.id)
                                    },
                     verticalArrangement = Arrangement.spacedBy(8.dp)
                 ) {
